cleaned font & colors
for lifts, trails and navigator bar.

+++ Weather.swift
@@ -26,7 +26,7 @@ class Weather {
     var currently = ""
     var currentTemperature: Float?
     var currentWindSpeed: Float?
-    var currentWindBearing: Float?
+    var currentWindBearing: Int?
     var precipType = "snow"
     var day0Snow: Float?
     var day1Snow: Float?
@@ -47,7 +47,7 @@ class Weather {
             if let windSpeed = currently[Weather.windSpeedKey] as? Float {
                 self.currentWindSpeed = windSpeed
             }
-            if let windBearing = currently[Weather.windBearingKey] as? Float {
+            if let windBearing = currently[Weather.windBearingKey] as? Int {
                 self.currentWindBearing = windBearing
             }
         }
